# Conflicts:
#	src/GUI/LandlordMakePaymentGUI.java

+++ src/Server/ReportHandler.java
@@ -17,8 +17,8 @@ public class ReportHandler extends Handler{
             return false;
         }
         String parts[] = input.split("#");
-        Date start = new Date(Integer.parseInt(parts[1]));
-        Date end  = new Date(Integer.parseInt(parts[2]));
+        Date start = new Date(Long.parseLong(parts[1]));
+        Date end  = new Date(Long.parseLong(parts[2]));
         String send = "";
         send+= "Number of total Properties: "+rpms.reportNumProperties()+"#";
